requirements.txt
streamlit==1.37.1
requests>=2.32.0
python-dotenv>=1.0.1
pandas>=2.0.0
openpyxl>=3.1.0
